requirements.txt
flask
torch
tifffile
numpy
pillow
segmentation-models-pytorch
opencv-python
